Move migrator and its interfaces to a subdirectory
