+++ examples/utils/migration-tools/src/migrator/index.ts
@@ -1,0 +1,14 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export {
+	DataTransformationCallback,
+	IMigrator,
+	IMigratorEvents,
+} from "./interfaces.js";
+export {
+	getModelAndMigrationToolFromContainer,
+	Migrator,
+} from "./migrator.js";
